Disable trusted domain warning
Because of the different subdomains for every sandstorm session it's not possible to add the host to the trusted domain list.
Due to the random subdomains the trusted domain provides no additional security when running on sandstorm.
This works for now, but there still might be a better way to disable trusted domains